requirements.txt
pillow==10.2.0
Flask==2.2.3